refactor: comment 로직 수정
로그인 인증 추가하여 comment 로직 수정
로그인한 사람이 작성한 comment 전체 조회 (RFP에 언급 x, 삭제할지 말지 논의)

+++ src/main/java/com/core/toy3/src/like/service/LikeService.java
@@ -51,7 +51,7 @@ public class LikeService {
     @Transactional
     public Response<List<LikeResponse>> getLikedTravelsByMemberId(Long memberId) {
         //memberId에 해당하는 Member를 조회
-        Member member = memberRepository.findById(memberId.intValue())
+        Member member = memberRepository.findById(memberId)
                 .orElseThrow(() -> new RuntimeException("Member not found with ID: " + memberId));
 
         // Member가 좋아요를 누른 목록을 조회
